Created LinkedList class and methods
I created my own version of Linked Lists in Ruby with some of the
essential methods to manipulate the lists.
I wrote two versions of most of the methods, one relying on the size()
method, the other without this reliance. I will probably opt for the
later one in later commits.